tagger/requirements.txt
numpy==1.26.4
pandas
onnxruntime-gpu==1.19.2
pillow
huggingface_hub
websockets

--extra-index-url https://download.pytorch.org/whl/cu124
torch==2.5.1